:white_check_mark: add tests and fix bug and seperate en file
