+++ thyrel-web/src/index.tsx
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom';
-import './i18n';
+import './i18n/i18n';
 import App from './App';
 import './index.css';
